pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pneuma-world"
version = "0.1.0"
description = "Multi-character autonomous world engine"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "pneuma-core>=0.1.0",
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pneuma_world"]
